feat: 에러 발생 시 슬랙 알람을 비동기로 보내도록 구현 (#69)
