+++ src/main/java/mocacong/server/controller/ControllerAdvice.java
@@ -1,19 +1,12 @@
 package mocacong.server.controller;
 
-import com.slack.api.Slack;
-import com.slack.api.model.Attachment;
-import com.slack.api.model.Field;
-import static com.slack.api.webhook.WebhookPayloads.payload;
-import java.io.IOException;
-import java.time.LocalDateTime;
-import java.time.format.DateTimeFormatter;
-import java.util.List;
 import java.util.Objects;
 import javax.servlet.http.HttpServletRequest;
+import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import mocacong.server.dto.response.ErrorResponse;
 import mocacong.server.exception.MocacongException;
-import org.springframework.beans.factory.annotation.Value;
+import mocacong.server.support.SlackAlarmGenerator;
 import org.springframework.http.ResponseEntity;
 import org.springframework.http.converter.HttpMessageNotReadableException;
 import org.springframework.validation.FieldError;
@@ -24,16 +17,14 @@ import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.RestControllerAdvice;
 
 @Slf4j
+@RequiredArgsConstructor
 @RestControllerAdvice
 public class ControllerAdvice {
 
     private static final int FIELD_ERROR_CODE_INDEX = 0;
     private static final int FIELD_ERROR_MESSAGE_INDEX = 1;
 
-    private final Slack slackClient = Slack.getInstance();
-
-    @Value("${slack.webhook.url}")
-    private String webhookUrl;
+    private final SlackAlarmGenerator slackAlarmGenerator;
 
     @ExceptionHandler(MethodArgumentNotValidException.class)
     public ResponseEntity<ErrorResponse> handleInputFieldException(MethodArgumentNotValidException e) {
@@ -79,44 +70,8 @@ public class ControllerAdvice {
                 request.getRequestURI(),
                 e.getMessage()
         );
-        sendSlackAlertErrorLog(e, request);
+        slackAlarmGenerator.sendSlackAlertErrorLog(e, request);
         return ResponseEntity.internalServerError()
                 .body(new ErrorResponse(9999, "일시적으로 접속이 원활하지 않습니다. 모카콩 서비스 팀에 문의 부탁드립니다."));
-    }
-
-    private void sendSlackAlertErrorLog(Exception e, HttpServletRequest request) {
-        try {
-            slackClient.send(webhookUrl, payload(p -> p
-                    .text("서버 에러 발생! 백엔드 측의 빠른 확인 요망")
-                    .attachments(
-                            List.of(generateSlackAttachment(e, request))
-                    )
-            ));
-        } catch (IOException slackError) {
-            log.debug("Slack 통신과의 예외 발생");
-        }
-    }
-
-    private Attachment generateSlackAttachment(Exception e, HttpServletRequest request) {
-        String requestTime = DateTimeFormatter.ofPattern("yyyy-MM-dd HH:mm:ss.SSS").format(LocalDateTime.now());
-        String xffHeader = request.getHeader("X-FORWARDED-FOR");
-        return Attachment.builder()
-                .color("ff0000")
-                .title(requestTime + " 발생 에러 로그")
-                .fields(List.of(
-                                generateSlackField("Request IP", xffHeader == null ? request.getRemoteAddr() : xffHeader),
-                                generateSlackField("Request URL", request.getRequestURL() + " " + request.getMethod()),
-                                generateSlackField("Error Message", e.getMessage())
-                        )
-                )
-                .build();
-    }
-
-    private Field generateSlackField(String title, String value) {
-        return Field.builder()
-                .title(title)
-                .value(value)
-                .valueShortEnough(false)
-                .build();
     }
 }
